+++ packages/google_mlkit_barcode_scanning/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.10.0
+
+* Fix: Update build.gradle, support AGP 8.
+* Update README.
+
+## 0.9.0
+
+* fix: barcode Rect.zero if rect not found.
+
 ## 0.8.0
 
 * fix: boundingBox and cornerPoint.